Version 1.10 of the flyway database has been created (a table mapping offer reservations has been added). Reservation entity have been created.
- An entity containing data about a offer reservation has been created.
- A repository with an adapter for basic management of the created entity was created.
- Exception handling was added when a reservation exists and reservation not found.
- DTO was added for reading and writing reservations.

Light improvements
- Set the default time zone in the configuration file to Europe/Warsaw for conversion to and from JSON.
- Change the id type in sql (1.5, table containing company offers) from int to bigint

+++ backend/src/main/resources/db/migration/V1_5__init_company_offers_table.sql
@@ -1,7 +1,7 @@
 drop table if exists companies_services;
 
 create table company_offers(
-    id int primary key auto_increment,
+    id bigint primary key auto_increment,
     name varchar(255) not null,
     description varchar(2048) not null,
     price double not null,
